Cambios en edicion de imagen

+++ src/main/java/com/example/demo/services/ImagenServiceImpl.java
@@ -30,7 +30,7 @@ public class ImagenServiceImpl implements ImagenService {
 	private ProductoRepository pRepository;
 	@Autowired
 	private UsuarioRepository uRepository;
-	
+
 	@Value("${url.image.nousuario}")
 	private String noUsuarioUrl;
 	@Value("${url.image.noproducto}")
@@ -45,7 +45,7 @@ public class ImagenServiceImpl implements ImagenService {
 		try {
 
 			Usuario usuario = uRepository.findById(idU).get();
-			Imagen imagen = iRepository.findByImagenUrl(usuario.getImage());
+			Imagen imagen = iRepository.findByUsuarioId(usuario.getId());
 			Map result = service.upload(file);
 			guardarImagen(usuario.getImage(), noUsuarioId, file, result, imagen, usuario.getId(), true);
 			usuario.setImage(result.get("url").toString());
@@ -59,31 +59,12 @@ public class ImagenServiceImpl implements ImagenService {
 	}
 
 	@Override
-	public RespuestaDto eliminarIU(String id) {
-		String msg = "Usuario no tine imagen";
-		try {
-			Usuario usuario = uRepository.findById(id).get();
-			if (!(usuario.getImage().contains(noUsuarioId))) {
-				msg = eliminarImagen(usuario.getImage(), "usuario");
-				usuario.setImage(noUsuarioUrl);
-				uRepository.save(usuario);
-			}
-			return new RespuestaDto(true, msg);
-		} catch (Exception e) {
-			log.error("Algo salio mal al eliminar la imagen del usuario" + e.getMessage());
-			return new RespuestaDto(false,
-					String.format("Algo salio mal, posiblemente el id: %s, del usuario no esta en la BD"
-							+ " ó la imagen ya fue eliminada", id));
-		}
-	}
-
-	@Override
 	public RespuestaDto guardarIP(MultipartFile file, String idP) {
 		try {
 			Map result = service.upload(file);
 
 			Producto producto = pRepository.findById(idP).get();
-			Imagen imagen = iRepository.findByImagenUrl(producto.getImagen());
+			Imagen imagen = iRepository.findByProductoId(producto.getId());
 			guardarImagen(producto.getImagen(), noProductoId, file, result, imagen, producto.getId(), false);
 			producto.setImagen(result.get("url").toString());
 			pRepository.save(producto);
@@ -92,25 +73,6 @@ public class ImagenServiceImpl implements ImagenService {
 			log.error("Algo salio mal al guardar la imagen del producto" + e.getMessage());
 			return new RespuestaDto(false,
 					String.format("Algo salio mal, posiblemente el id: %s, del usuario no esta en la BD", idP));
-		}
-	}
-
-	@Override
-	public RespuestaDto eliminarIP(String id) {
-		String msg = "Producto no tine imagen";
-		try {
-			Producto producto = pRepository.findById(id).get();
-			if (!(producto.getImagen().contains("goev2g0h9li7wbcajaet"))) {
-				msg = eliminarImagen(producto.getImagen(), "producto");
-				producto.setImagen(noProductoUrl);
-				pRepository.save(producto);
-			}
-			return new RespuestaDto(true, msg);
-		} catch (Exception e) {
-			log.error("Algo salio mal al eliminar la imagen del producto" + e.getMessage());
-			return new RespuestaDto(false,
-					String.format("Algo salio mal, posiblemente el id: %s, del producto no esta en la BD"
-							+ " ó la imagen ya fue eliminada", id));
 		}
 	}
 
@@ -133,10 +95,56 @@ public class ImagenServiceImpl implements ImagenService {
 		}
 	}
 
-	private String eliminarImagen(final String urlImage, String msg) throws IOException {
-		Imagen imagen = iRepository.findByImagenUrl(urlImage);
-		service.delete(imagen.getImagenId());
-		iRepository.delete(imagen);
-		return String.format("Imagen de %s eliminada", msg);
+	@Override
+	public RespuestaDto eliminarI(String id) {
+		try {
+			service.delete(id);
+			return new RespuestaDto(true, "Eliminación de la imagen sin problema");
+		} catch (Exception e) {
+			log.error("Error al eliminar la imagen ", e.getLocalizedMessage());
+			return new RespuestaDto(false, e.getMessage());
+		}
+	}
+
+	@Override
+	public RespuestaDto editarIP(Producto producto,String id) {
+		Imagen imagen = iRepository.findByProductoId(id);
+		if(imagen != null) {
+		try {
+			if (!(producto.getImagen().contains(noProductoId)))
+				service.delete(imagen.getImagenId());
+		} catch (IOException e) {
+			log.error("Algo salio mal al eliminar la imagen del producto");
+			return new RespuestaDto(false,"Algo salio mal al eliminar la imagen del producto");
+		}
+		imagen.setImagenId(producto.getIdImagen());
+		imagen.setImagenUrl(producto.getImagen());
+		iRepository.save(imagen);
+		pRepository.save(producto);
+		return new RespuestaDto(true,String.format("Imagen de %s actualizada",producto.getNombre()));
+	}else {
+		return new RespuestaDto(false,String.format("Id:%s de Producto NO existe en BD"));
+	}
+	}
+
+	@Override
+	public RespuestaDto editarIU(Usuario usuario, String id) {
+		Imagen imagen = iRepository.findByUsuarioId(id);
+		if(imagen != null) {
+			try {
+				if (!(usuario.getImage().contains(noUsuarioId)))
+					service.delete(imagen.getImagenId());
+			} catch (IOException e) {
+				log.error("Algo salio mal al eliminar la imagen del producto");
+				return new RespuestaDto(false,"Algo salio mal al eliminar la imagen del producto");
+			}
+			imagen.setImagenId(usuario.getIdImage());
+			imagen.setImagenUrl(usuario.getImage());
+			iRepository.save(imagen);
+			uRepository.save(usuario);
+			return new RespuestaDto(true,String.format("Imagen de %s actualizada",usuario.getName()));
+		}else {
+			return new RespuestaDto(false,String.format("Id:%s de usuario NO existe en BD"));
+		}
 	}
 }
